tools: Makefile: rework the objcopy logic to embeded the dtb
When compiling with recent arm-none-eabi toolchain, there is a conflict
between the EABI version on the device tree object create by objcopy
(EABI version 0) and the toolchain EABI version (EABI version 5).

Objcopy seems unable to create object with other EABI version than 0.

A fix is to use the .inc directive in an assembly file to embedded the
dtb. This way, the assembly file is compiling use AS.